backend/pyproject.toml
[tool.poetry]
name = "doit"
version = "0.1.0"
description = ""
authors = ["riisatoro <riisa.toro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.1"
psycopg2-binary = "^2.9.1"
requests = "^2.25.1"
html2text = "^2020.1.16"
itsdangerous = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
